fix(ipc): migrate clear-quarantined-panel to defineIpcNamespace
CI ratchet (scripts/ipc-handwritten-ratchet.mjs) requires new IPC
channels to flow through defineIpcNamespace + codegen rather than
hand-written entries in maps.ts. Move the handler into a new
electron/ipc/handlers/safeMode.ts module using defineIpcNamespace +
op(), wire it into the registration list, regenerate generated.ts,
and remove the hand-written maps.ts entry. Handler test moved to
electron/ipc/__tests__/safeMode.test.ts.

No behavior change — same channel name, same payload shape.
