+++ electron/ipc/__tests__/app.state.test.ts
@@ -655,47 +655,3 @@ describe("app:boot handler", () => {
     expect(projectStore.getCurrentProject).not.toHaveBeenCalled();
   });
 });
-
-describe("app:clear-quarantined-panel handler", () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-    ipcHandlers.clear();
-    panelSuspectLedger.clearQuarantinedPanel.mockReset();
-  });
-
-  async function invokeClear(panelId: unknown) {
-    const cleanup = registerAppStateHandlers();
-    const handler = ipcHandlers.get("app:clear-quarantined-panel");
-    if (!handler) throw new Error("app:clear-quarantined-panel handler not registered");
-    // ipcMain.handle passes (event, ...args) — the test's handler captures the
-    // identity of the wrapped callback, so we must reproduce the IPC-event arg.
-    const result = (await handler({}, panelId)) as { cleared: boolean };
-    cleanup();
-    return result;
-  }
-
-  it("forwards the request to the ledger when given a non-empty string id", async () => {
-    panelSuspectLedger.clearQuarantinedPanel.mockReturnValue(true);
-    const result = await invokeClear("panel-1");
-    expect(panelSuspectLedger.clearQuarantinedPanel).toHaveBeenCalledWith("panel-1");
-    expect(result).toEqual({ cleared: true });
-  });
-
-  it("reports cleared=false when the ledger had no record", async () => {
-    panelSuspectLedger.clearQuarantinedPanel.mockReturnValue(false);
-    const result = await invokeClear("missing-panel");
-    expect(result).toEqual({ cleared: false });
-  });
-
-  it("rejects an empty string id without calling the ledger", async () => {
-    const result = await invokeClear("");
-    expect(panelSuspectLedger.clearQuarantinedPanel).not.toHaveBeenCalled();
-    expect(result).toEqual({ cleared: false });
-  });
-
-  it("rejects non-string ids without calling the ledger", async () => {
-    const result = await invokeClear(42);
-    expect(panelSuspectLedger.clearQuarantinedPanel).not.toHaveBeenCalled();
-    expect(result).toEqual({ cleared: false });
-  });
-});
